Restructured the code as per cli repo
