--- cmd/ocm-support/accounts/cmd.go
@@ -1,19 +1,0 @@
-package accounts
-
-import (
-	"github.com/spf13/cobra"
-
-	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/accounts/find"
-)
-
-// Cmd ...
-var Cmd = &cobra.Command{
-	Use:   "accounts COMMAND",
-	Short: "Gets information about or execute actions on accounts.",
-	Long:  "Gets information about or execute actions on accounts.",
-	Args:  cobra.MinimumNArgs(1),
-}
-
-func init() {
-	Cmd.AddCommand(find.Cmd)
-}
