+++ cmd/ocm-support/create/cmd.go
@@ -3,24 +3,29 @@ package create
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/account"
-	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/organization"
-	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/subscription"
+	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/create/account_capability"
+	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/create/account_label"
+	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/create/organization_capability"
+	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/create/organization_label"
+	registry_credentials "github.com/openshift-online/ocm-support-cli/cmd/ocm-support/create/registryCredentials"
+	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/create/subscription_capability"
+	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/create/subscription_label"
 )
 
 // Cmd ...
 var Cmd = &cobra.Command{
 	Use:   "create [COMMAND]",
-	Short: "Assigns Labels, Capabilities to Accounts, Subscriptions, Organizations",
-	Long:  "Assigns Labels, Capabilities to Accounts, Subscriptions, Organizations",
+	Short: "Creates the given resource",
+	Long:  "Creates the given resource",
 	Args:  cobra.MinimumNArgs(1),
 }
 
 func init() {
-	Cmd.AddCommand(account.CmdCreateAccountLabel)
-	Cmd.AddCommand(organization.CmdCreateOrganizationLabel)
-	Cmd.AddCommand(subscription.CmdCreateSubscriptionLabel)
-	Cmd.AddCommand(account.CmdCreateAccountCapability)
-	Cmd.AddCommand(organization.CmdCreateOrganizationCapability)
-	Cmd.AddCommand(subscription.CmdCreateSubscriptionCapability)
+	Cmd.AddCommand(account_label.CmdCreateAccountLabel)
+	Cmd.AddCommand(organization_label.CmdCreateOrganizationLabel)
+	Cmd.AddCommand(subscription_label.CmdCreateSubscriptionLabel)
+	Cmd.AddCommand(account_capability.CmdCreateAccountCapability)
+	Cmd.AddCommand(organization_capability.CmdCreateOrganizationCapability)
+	Cmd.AddCommand(subscription_capability.CmdCreateSubscriptionCapability)
+	Cmd.AddCommand(registry_credentials.CmdCreateRegistryCredentials)
 }
