+++ cmd/ocm-support/create/organization_capability/cmd.go
@@ -1,4 +1,4 @@
-package account
+package organization_capability
 
 import (
 	"fmt"
@@ -8,32 +8,32 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/utils"
-	"github.com/openshift-online/ocm-support-cli/pkg/account"
 	"github.com/openshift-online/ocm-support-cli/pkg/capability"
 	"github.com/openshift-online/ocm-support-cli/pkg/label"
+	"github.com/openshift-online/ocm-support-cli/pkg/organization"
 )
 
-// CmdCreateAccountCapability represents the create account capability command
-var CmdCreateAccountCapability = &cobra.Command{
-	Use:   "accountCapability [accountID] [capability]",
-	Short: "Assigns a Capability to an Account",
-	Long:  "Assigns a Capability to an Account",
-	RunE:  runCreateAccountCapability,
+// CmdCreateOrganizationCapability represents the create organization capability command
+var CmdCreateOrganizationCapability = &cobra.Command{
+	Use:   "organizationCapability [organizationID] [capability]",
+	Short: "Assigns a Capability to an Organization",
+	Long:  "Assigns a Capability to an Organization",
+	RunE:  runCreateOrganizationCapability,
 	Args:  cobra.ExactArgs(2),
 	PreRunE: func(cmd *cobra.Command, args []string) error {
-		accountID := args[0]
+		organizationID := args[0]
 		connection, err := ocm.NewConnection().Build()
 		if err != nil {
 			return fmt.Errorf("failed to create OCM connection: %v", err)
 		}
-		// validates the account
-		err = account.ValidateAccount(accountID, connection)
+		// validates the organization
+		err = organization.ValidateOrganization(organizationID, connection)
 		if err != nil {
 			return fmt.Errorf("%v", err)
 		}
 		//validates the capability
 		capabilityKey := args[1]
-		err = capability.ValidateCapability(capabilityKey, "account")
+		err = capability.ValidateCapability(capabilityKey, "organization")
 		if err != nil {
 			return fmt.Errorf("%v", err)
 		}
@@ -41,19 +41,19 @@ var CmdCreateAccountCapability = &cobra.Command{
 	},
 }
 
-func runCreateAccountCapability(cmd *cobra.Command, argv []string) error {
-	accountID := argv[0]
+func runCreateOrganizationCapability(cmd *cobra.Command, argv []string) error {
+	organizationID := argv[0]
 	// TODO : avoid creating multiple connection pools
 	connection, err := ocm.NewConnection().Build()
 	if err != nil {
 		return fmt.Errorf("failed to create OCM connection: %v", err)
 	}
 	key := argv[1]
-	capabilityKey, err := capability.GetCapability(key, "account")
+	capabilityKey, err := capability.GetCapability(key, "organization")
 	if err != nil {
 		return fmt.Errorf("failed to get capability: %v", err)
 	}
-	createdCapability, err := account.AddLabel(accountID, capabilityKey, "true", true, connection)
+	createdCapability, err := organization.AddLabel(organizationID, capabilityKey, "true", true, connection)
 	if err != nil {
 		return fmt.Errorf("failed to create capability: %v", err)
 	}
