+++ cmd/ocm-support/create/subscription_capability/cmd.go
@@ -1,4 +1,4 @@
-package subscription
+package subscription_capability
 
 import (
 	"fmt"
